Save docker name in vnfci.Hostname
Currently hostname stored in vnfci is vnfd.name. When docker instantiates this VDU it automatically provides a
name by adding couple of digits of random number to it. This patches updates vnfci.Hostname with the actual
docker instance name.

This change is necessary to integrate zabbix based monitoring of docker instances.